requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# AWS SDK
boto3==1.34.84
botocore==1.34.84

# Google Cloud SDK
google-cloud-storage==2.13.0

# Validation and Serialization
marshmallow==3.20.2

# Development and Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
requests==2.31.0

# Production WSGI Server
gunicorn==21.2.0

# Monitoring and Logging
structlog==23.2.0
